Introduced ConfManager for conf related codes
