+++ smart-server/src/main/java/org/smartdata/server/engine/ConfManager.java
@@ -1,0 +1,33 @@
+package org.smartdata.server.engine;
+
+import org.smartdata.conf.ReconfigurableBase;
+import org.smartdata.conf.ReconfigureException;
+import org.smartdata.conf.SmartConf;
+import org.smartdata.conf.SmartConfKeys;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class ConfManager extends ReconfigurableBase {
+
+  private SmartConf conf;
+
+  public ConfManager(SmartConf conf) {
+    this.conf = conf;
+  }
+
+  @Override
+  public void reconfigureProperty(String property, String newVal)
+      throws ReconfigureException {
+    if (property.equals(SmartConfKeys.DFS_SSM_NAMENODE_RPCSERVER_KEY)) {
+      conf.set(property, newVal);
+    }
+  }
+
+  @Override
+  public List<String> getReconfigurableProperties() {
+    return Arrays.asList(
+        SmartConfKeys.DFS_SSM_NAMENODE_RPCSERVER_KEY
+    );
+  }
+}
